Add the negation `!` form: `! cmd1 | cmd2`
Allow a lone `!` to be a vaild command - this is not specified by POSIX,
but some shells support it.

Allow using multiple `!`s in sequence - again, not specified by POSIX
but bash seems to support it.

A lone `!` sets $? to 1, but a lone `! !` doesn't change $?. This
'feature' is probably used by no one, because `true` and `false` exist.